+++ unittest/UMEUnitTestSimd128b.h
@@ -971,14 +971,14 @@ int test_UME_SIMD4_32u(bool supressMessages)
     }
     {
         alignas(16) uint32_t arr[10] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32u vec0(1);
         vec0.gather(arr, indices);
         CHECK_CONDITION(vec0[0] == 2 && vec0[1] == 4 && vec0[2] == 9 && vec0[3] == 6, "GATHERS");
     }
     {
         alignas(16) uint32_t arr[10] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32u vec0(1);
         SIMDMask4 mask(true, false, true, false);
         vec0.gather(mask, arr, indices);
@@ -1001,7 +1001,7 @@ int test_UME_SIMD4_32u(bool supressMessages)
     }
     {
         alignas(16) uint32_t arr[10] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32u vec0(9, 8, 7, 6);
         vec0.scatter(arr, indices);
         CHECK_CONDITION(arr[1] == 9 && arr[3] == 8 && arr[8] == 7 && arr[9] == 10, "SCATTERS");
@@ -1570,14 +1570,14 @@ int test_UME_SIMD4_32i(bool supressMessages)
     }
     {
         alignas(16) int32_t arr[10] = {1, -2, 3, 4, 5, -6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32i vec0(1);
         vec0.gather(arr, indices);
         CHECK_CONDITION(vec0[0] == -2 && vec0[1] == 4 && vec0[2] == 9 && vec0[3] == -6, "GATHERS");
     }
     {
         alignas(16) int32_t arr[10] = {1, -2, 3, 4, 5, -6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32i vec0(1);
         SIMDMask4 mask(true, false, true, false);
         vec0.gather(mask, arr, indices);
@@ -1600,7 +1600,7 @@ int test_UME_SIMD4_32i(bool supressMessages)
     }
     {
         alignas(16) int32_t arr[10] = {1, -2, 3, 4, 5, -6, 7, 8, 9, 10};
-        uint64_t indices[] = {1, 3, 8, 5};
+        uint32_t indices[] = {1, 3, 8, 5};
         SIMD4_32i vec0(9, -8, 7, 6);
         vec0.scatter(arr, indices);
         CHECK_CONDITION(arr[1] == 9 && arr[3] == -8 && arr[8] == 7 && arr[9] == 10, "SCATTERS");
